Remove FIR and ConcurrentQueue from SystemC_test.cpp
