+++ sparta/test/SystemC/CMakeLists.txt
@@ -4,7 +4,7 @@ if (NOT DEFINED ENV{SYSTEMC_HOME})
   message (FATAL_ERROR "SYSTEMC_HOME is not defined -- cannot include this test")
 endif ()
 
-sparta_add_test_executable(SystemC_test SystemC_test.cpp fir.cpp tb.cpp)
+sparta_add_test_executable(SystemC_test SystemC_test.cpp)
 target_link_libraries     (SystemC_test SystemC::systemc)
 
 include(../TestingMacros.cmake)
